timeSeriesService/requirements.txt
pika
requests
python-dateutil